arch-nspawn: limit the machine_name length to ${HOST_NAME_MAX:-64} chars.
The criteria for a valid name are:
 - is made of of alphanumerics, "_", ".", and "-"
 - is not emtpy
 - cannot end with "."
 - cannot contain ".."
 - the maximum length is HOST_NAME_MAX, as defined in <limits.h>

The existing code takes care of all but the last of these.  So, added code
after the existing code to check the length of the string, and grab the
longest possible substring from the end.

I hard-coded our interpretation of HOST_NAME_MAX as 64--on Linux that has
been the value since version 1.0 of the kernel... I don't think it will be
changing any time soon, and since systemd is Linux-only, we don't have to
worry about other kernels.

But, just in case, I included a commented out version that uses cpp to get
the value.