arch-nspawn: Message style: two spaces after a period
This affects both the usage() text, and the error message if the
`/.arch-chroot` version doesn't match.  The latter is the one that I
really care about, and motivates this change.

On Parabola, the `arch-nspawn` program isn't in PATH, it's somewhere
under `/usr/lib/`, and gets called as a helper to user-facing
programs; and the error message is displayed directly to the user.
These programs consistently put two spaces after a period when
printing a message to the terminal.